requirements.txt
streamlit
google-generativeai
pandas
numpy
plotly
gTTS
python-dotenv